+++ exp/tools/dump-ledger-state/diff_test.sh
@@ -8,7 +8,7 @@ do
     continue
   fi
   wc -l ${i}.csv
-  sort -o ${i}_sorted.csv ${i}.csv
+  sort -S 1G -o ${i}_sorted.csv ${i}.csv
 done
 
 echo "Sorting stellar-core output files..."
@@ -19,17 +19,18 @@ do
     continue
   fi
   wc -l ${i}_core.csv
-  sort -o ${i}_core_sorted.csv ${i}_core.csv
+  sort -S 1G -o ${i}_core_sorted.csv ${i}_core.csv
 done
 
 echo "Checking diffs..."
-for i in "${ENTRIES[@]}"
+for type in "${ENTRIES[@]}"
 do
-  diff -q ${i}_core_sorted.csv ${i}_sorted.csv
+  diff -q ${type}_core_sorted.csv ${type}_sorted.csv
   if [ "$?" -ne "0" ]
   then
-    echo "ERROR: $i does NOT match";
+    echo "ERROR: $type does NOT match";
+    exit -1
   else
-    echo "$i OK";
+    echo "$type OK";
   fi
 done