+++ services/horizon/internal/test/scenarios/base-horizon.sql
@@ -529,6 +529,8 @@ INSERT INTO gorp_migrations VALUES ('14_fix_asset_toml_field.sql', '2019-06-03 1
 INSERT INTO gorp_migrations VALUES ('15_ledger_failed_txs.sql', '2019-06-03 18:28:47.116796+02');
 INSERT INTO gorp_migrations VALUES ('16_ingest_failed_transactions.sql', '2019-06-03 18:28:47.117989+02');
 INSERT INTO gorp_migrations VALUES ('17_transaction_fee_paid.sql', '2019-06-03 18:28:47.120034+02');
+INSERT INTO gorp_migrations VALUES ('18_account_for_signers.sql', '2019-06-03 18:28:47.120034+02');
+INSERT INTO gorp_migrations VALUES ('19_offers.sql', '2019-06-03 18:28:47.120034+02');
 
 
 --